Split up SmallSceneryPlace/Remove
Added undo function for Remove Scenery
