+++ src/openrct2-ui/interface/ViewportInteraction.cpp
@@ -22,7 +22,7 @@
 #include <openrct2/actions/FootpathAdditionRemoveAction.hpp>
 #include <openrct2/actions/LargeSceneryRemoveAction.hpp>
 #include <openrct2/actions/ParkEntranceRemoveAction.hpp>
-#include <openrct2/actions/SmallSceneryRemoveAction.hpp>
+#include <openrct2/actions/SmallSceneryRemoveAction.h>
 #include <openrct2/actions/WallRemoveAction.hpp>
 #include <openrct2/localisation/Localisation.h>
 #include <openrct2/ride/Ride.h>
