+++ src/openrct2/actions/GameActionRegistration.cpp
@@ -66,8 +66,8 @@
 #include "SetParkEntranceFeeAction.hpp"
 #include "SignSetNameAction.hpp"
 #include "SignSetStyleAction.hpp"
-#include "SmallSceneryPlaceAction.hpp"
-#include "SmallSceneryRemoveAction.hpp"
+#include "SmallSceneryPlaceAction.h"
+#include "SmallSceneryRemoveAction.h"
 #include "SmallScenerySetColourAction.hpp"
 #include "StaffFireAction.hpp"
 #include "StaffHireNewAction.hpp"
