+++ src/openrct2/ride/TrackDesign.cpp
@@ -23,8 +23,8 @@
 #include "../actions/RideSetName.hpp"
 #include "../actions/RideSetSetting.hpp"
 #include "../actions/RideSetVehiclesAction.hpp"
-#include "../actions/SmallSceneryPlaceAction.hpp"
-#include "../actions/SmallSceneryRemoveAction.hpp"
+#include "../actions/SmallSceneryPlaceAction.h"
+#include "../actions/SmallSceneryRemoveAction.h"
 #include "../actions/TrackPlaceAction.hpp"
 #include "../actions/TrackRemoveAction.hpp"
 #include "../actions/WallPlaceAction.hpp"
